Improve Bazel XCHammer consumption / distribution
This PR makes it easier to pull XCHammer into a WORKSPACE with Bazel. The
possibilities are:

*xchammer_dist_repo*
Pulled as a binary distributed artifact e.g. github release

*xchammer_dev_repo*
From a local repository, but built inside of the respective repository

*source: git_repository or local_repository*
pulling in the entire repository and building it in tree

Additionally includes multiple breaking changes to the public starlark
API
- it has to rename `xchammer_resources` to `xchammer`
- it removes the need to specify xchammer in the `xcode_project` rule.
This invariant is enforced by the fact the binary is always in the same
repo as the rules.